Consolidate test fixtures
