+++ tests/blockchain/test_blockchain_transactions.py
@@ -11,9 +11,9 @@ from chia.types.spend_bundle import SpendBundle
 from chia.util.errors import ConsensusError, Err
 from chia.util.ints import uint64
 from tests.blockchain.blockchain_test_utils import _validate_and_add_block
-from tests.wallet_tools import WalletTool
 from tests.setup_nodes import test_constants
 from tests.util.generator_tools_testing import run_and_get_removals_and_additions
+from tests.wallet_tools import WalletTool
 
 BURN_PUZZLE_HASH = b"0" * 32
