+++ tests/core/full_node/full_sync/test_full_sync.py
@@ -17,7 +17,6 @@ from tests.core.node_height import node_height_exactly, node_height_between
 from tests.setup_nodes import test_constants
 from tests.time_out_assert import time_out_assert
 
-
 log = logging.getLogger(__name__)
