+++ tests/core/test_filter.py
@@ -1,16 +1,7 @@
 from typing import List
 
 import pytest
-import pytest_asyncio
 from chiabip158 import PyBIP158
-
-from tests.setup_nodes import setup_simulators_and_wallets
-
-
-@pytest_asyncio.fixture(scope="function")
-async def wallet_and_node():
-    async for _ in setup_simulators_and_wallets(1, 1, {}):
-        yield _
 
 
 class TestFilter:
