+++ tests/core/test_full_node_rpc.py
@@ -1,9 +1,7 @@
 # flake8: noqa: F811, F401
-import logging
 from typing import List
 
 import pytest
-import pytest_asyncio
 from blspy import AugSchemeMPL
 
 from chia.consensus.pot_iterations import is_overflow_block
@@ -16,22 +14,16 @@ from chia.simulator.simulator_protocol import FarmNewBlockProtocol, ReorgProtoco
 from chia.types.full_block import FullBlock
 from chia.types.spend_bundle import SpendBundle
 from chia.types.unfinished_block import UnfinishedBlock
-from tests.block_tools import get_signage_point
 from chia.util.hash import std_hash
-from chia.util.ints import uint16, uint8
+from chia.util.ints import uint8
+from tests.block_tools import get_signage_point
 from tests.blockchain.blockchain_test_utils import _validate_and_add_block
-from tests.wallet_tools import WalletTool
 from tests.connection_utils import connect_and_get_peer
-from tests.setup_nodes import setup_simulators_and_wallets, test_constants
+from tests.setup_nodes import test_constants
 from tests.time_out_assert import time_out_assert
 from tests.util.rpc import validate_get_routes
 from tests.util.socket import find_available_listen_port
-
-
-@pytest_asyncio.fixture(scope="function")
-async def two_nodes_sim_and_wallets():
-    async for _ in setup_simulators_and_wallets(2, 0, {}):
-        yield _
+from tests.wallet_tools import WalletTool
 
 
 class TestRpc:
