+++ tests/wallet/rl_wallet/test_rl_wallet.py
@@ -1,18 +1,10 @@
 import pytest
-import pytest_asyncio
 
 from chia.simulator.simulator_protocol import FarmNewBlockProtocol
 from chia.types.peer_info import PeerInfo
 from chia.util.ints import uint16, uint64
 from chia.wallet.rl_wallet.rl_wallet import RLWallet
-from tests.setup_nodes import setup_simulators_and_wallets
 from tests.time_out_assert import time_out_assert
-
-
-@pytest_asyncio.fixture(scope="function")
-async def two_wallet_nodes():
-    async for _ in setup_simulators_and_wallets(1, 2, {}):
-        yield _
 
 
 class TestCATWallet:
