+++ tests/wallet/sync/test_wallet_sync.py
@@ -2,7 +2,6 @@
 import asyncio
 
 import pytest
-import pytest_asyncio
 from colorlog import getLogger
 
 from chia.consensus.block_rewards import calculate_base_farmer_reward, calculate_pool_reward
@@ -13,7 +12,7 @@ from chia.util.ints import uint16, uint32
 from chia.wallet.wallet_state_manager import WalletStateManager
 from tests.connection_utils import disconnect_all_and_reconnect
 from tests.pools.test_pool_rpc import wallet_is_synced
-from tests.setup_nodes import setup_node_and_wallet, setup_simulators_and_wallets, test_constants
+from tests.setup_nodes import test_constants
 from tests.time_out_assert import time_out_assert
 
 
@@ -25,24 +24,6 @@ def wallet_height_at_least(wallet_node, h):
 
 
 log = getLogger(__name__)
-
-
-@pytest_asyncio.fixture(scope="function")
-async def wallet_node(self_hostname):
-    async for _ in setup_node_and_wallet(test_constants, self_hostname):
-        yield _
-
-
-@pytest_asyncio.fixture(scope="function")
-async def wallet_node_simulator():
-    async for _ in setup_simulators_and_wallets(1, 1, {}):
-        yield _
-
-
-@pytest_asyncio.fixture(scope="function")
-async def wallet_node_starting_height(self_hostname):
-    async for _ in setup_node_and_wallet(test_constants, self_hostname, starting_height=100):
-        yield _
 
 
 class TestWalletSync:
